Skip resources that don't exist

+++ concourse_janitor.py
@@ -33,36 +33,39 @@ def handler(event: None, context: None) -> None:  # pylint: disable=unused-argum
 
     log_group_response = logs.describe_log_groups()
 
-    for log_group in log_group_response["logGroups"]:
-        log_group_name = log_group["logGroupName"]
-        match_parts = search(INSTANCE_ID_REGEX, log_group_name)
-        if match_parts is not None:
-            instance_id = match_parts.group("instance_id")
-            if instance_id not in instance_ids:
-                logs.delete_log_group(logGroupName=log_group_name)
-                print(f"Deleted log group {log_group_name}")
+    if "logGroups" in log_group_response:
+        for log_group in log_group_response["logGroups"]:
+            log_group_name = log_group["logGroupName"]
+            match_parts = search(INSTANCE_ID_REGEX, log_group_name)
+            if match_parts is not None:
+                instance_id = match_parts.group("instance_id")
+                if instance_id not in instance_ids:
+                    logs.delete_log_group(logGroupName=log_group_name)
+                    print(f"Deleted log group {log_group_name}")
 
     sqs_response = sqs.list_queues()
 
-    for queue_url in sqs_response["QueueUrls"]:
-        match_parts = search(INSTANCE_ID_REGEX, queue_url)
-        if match_parts is not None:
-            instance_id = match_parts.group("instance_id")
-            if instance_id not in instance_ids:
-                sqs.delete_queue(QueueUrl=queue_url)
-                print(f"Deleted queue {queue_url}")
+    if "QueueUrls" in sqs_response:
+        for queue_url in sqs_response["QueueUrls"]:
+            match_parts = search(INSTANCE_ID_REGEX, queue_url)
+            if match_parts is not None:
+                instance_id = match_parts.group("instance_id")
+                if instance_id not in instance_ids:
+                    sqs.delete_queue(QueueUrl=queue_url)
+                    print(f"Deleted queue {queue_url}")
 
     route53_response = route53.list_resource_record_sets(HostedZoneId=HOSTED_ZONE_ID)
     route_53_changes = []
 
-    for resource_record_set in route53_response["ResourceRecordSets"]:
-        name = resource_record_set["Name"]
-        match_parts = search(INSTANCE_ID_REGEX, name)
-        if match_parts is not None:
-            instance_id = match_parts.group("instance_id")
-            if instance_id not in instance_ids:
-                route_53_changes.append({"Action": "DELETE", "ResourceRecordSet": resource_record_set})
-                print(f"Queued {name} for deletion")
+    if "ResourceRecordSets" in route53_response:
+        for resource_record_set in route53_response["ResourceRecordSets"]:
+            name = resource_record_set["Name"]
+            match_parts = search(INSTANCE_ID_REGEX, name)
+            if match_parts is not None:
+                instance_id = match_parts.group("instance_id")
+                if instance_id not in instance_ids:
+                    route_53_changes.append({"Action": "DELETE", "ResourceRecordSet": resource_record_set})
+                    print(f"Queued {name} for deletion")
 
     if len(route_53_changes) > 0:
         route53.change_resource_record_sets(
